# services/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.CreateTenantRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
      plan:
        type: string
      slug:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.InviteMemberRequest:
    properties:
      role:
        type: string
      user_id:
        type: string
    required:
    - user_id
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      session_id:
        type: integer
      user_id:
        type: string
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.LogoutRequest:
    properties:
      session_id:
        type: integer
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RegisterResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      session_id:
        type: integer
      user_id:
        type: string
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.TenantMemberResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        type: string
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.TenantResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      plan:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.UpdateTenantRequest:
    properties:
      name:
        type: string
      plan:
        type: string
      status:
        type: string
    type: object
  ? github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-array_github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantMemberResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.TenantMemberResponse'
        type: array
      meta: {}
    type: object
  ? github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-array_github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.TenantResponse'
        type: array
      meta: {}
    type: object
  ? github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantMemberResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.TenantMemberResponse'
      meta: {}
    type: object
  ? github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.TenantResponse'
      meta: {}
    type: object
host: localhost:9090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for CommerceHub backend service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CommerceHub API
  version: "1.0"
paths:
  /api/v1/auth/google/callback:
    get:
      description: Handles Google OAuth callback, validates state, and redirects to
        frontend with secure cookies
      parameters:
      - description: OAuth code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth CSRF state
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
      summary: Google OAuth callback
      tags:
      - Auth
  /api/v1/auth/google/login:
    get:
      description: Redirects to Google consent screen with secure CSRF state
      responses:
        "302":
          description: Found
      summary: Google OAuth login
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.LoginResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the user's current or specified session (enforcing ownership)
      parameters:
      - description: Logout request
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RefreshTokenResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Register credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.RegisterResponse'
      summary: Register user
      tags:
      - Auth
  /api/v1/tenants:
    get:
      consumes:
      - application/json
      description: Get all workspaces the authenticated user belongs to
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-array_github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantResponse'
      security:
      - BearerAuth: []
      summary: List user workspaces
      tags:
      - Tenants
    post:
      consumes:
      - application/json
      description: Create a new multi-tenant workspace
      parameters:
      - description: Workspace data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantResponse'
      security:
      - BearerAuth: []
      summary: Create workspace
      tags:
      - Tenants
  /api/v1/tenants/{id}:
    get:
      consumes:
      - application/json
      description: Get details for a specific workspace (caller must be a member)
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantResponse'
      security:
      - BearerAuth: []
      summary: Get workspace details
      tags:
      - Tenants
    put:
      consumes:
      - application/json
      description: Update workspace details (requires owner or admin role)
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Workspace update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantResponse'
      security:
      - BearerAuth: []
      summary: Update workspace settings
      tags:
      - Tenants
  /api/v1/tenants/{id}/invites:
    post:
      consumes:
      - application/json
      description: Invite a user to join the workspace (requires owner or admin role)
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantMemberResponse'
      security:
      - BearerAuth: []
      summary: Invite user to workspace
      tags:
      - Tenants
  /api/v1/tenants/{id}/users:
    get:
      consumes:
      - application/json
      description: Get all members belonging to a workspace
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TruongHoang2004_Hexta_services_api_internal_present_http_response.Response-array_github_com_TruongHoang2004_Hexta_services_api_internal_present_http_dto_TenantMemberResponse'
      security:
      - BearerAuth: []
      summary: List workspace members
      tags:
      - Tenants
  /health:
    get:
      consumes:
      - application/json
      description: Check the health of the application and its dependencies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health Check
      tags:
      - System
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"